chore: move modal logic to utils
